feat: add support for labels in func.yaml
This change adds support for setting labels on deployed functions. It does
not include any command line ability to do so. It was unclear how to best
test this, as YAML marshaling will blow up if that bit is incorrect.

I suppose the ideal test would be to actually add a label in an integration
test and check to see if it's there on deployment.

Signed-off-by: Lance Ball <lball@redhat.com>